Avoid breaking sugar items throughout the run
