+++ src/engine/outfit.ts
@@ -1,4 +1,4 @@
-import { cliExecute, equip, equippedItem } from "kolmafia";
+import { cliExecute, equip, equippedItem, Item } from "kolmafia";
 import { $item, $skill, $slot, get, have } from "libram";
 
 export function garbageShirt(): void {
@@ -29,4 +29,18 @@ export function docBag(): void {
     !get("instant_saveDocBag", false)
   )
     equip($slot`acc3`, $item`Lil' Doctor™ bag`);
+}
+
+export function sugarItemsAboutToBreak(): Item[] {
+  const sugarItems = [
+    { id: 4181, item: $item`sugar chapeau` },
+    { id: 4182, item: $item`sugar shorts` },
+  ];
+  return sugarItems
+    .map((entry) => {
+      const { id, item } = entry;
+      const itemAboutToBreak = parseInt(get(`sugarCounter${id.toString()}`), 10) >= 30;
+      return itemAboutToBreak ? [item] : [];
+    })
+    .reduce((a, b) => a.concat(b));
 }
