+++ src/tasks/spelldamage.ts
@@ -23,6 +23,7 @@ import {
 import { Quest } from "../engine/task";
 import { advCost, CommunityServiceTests, logTestSetup, tryAcquiringEffect } from "../lib";
 import Macro from "../combat";
+import { sugarItemsAboutToBreak } from "../engine/outfit";
 
 export const SpellDamageQuest: Quest = {
   name: "Spell Damage",
@@ -47,10 +48,11 @@ export const SpellDamageQuest: Quest = {
           .trySkill($skill`Use the Force`)
           .abort()
       ),
-      outfit: {
+      outfit: () => ({
         weapon: $item`Fourth of May Cosplay Saber`,
         familiar: $familiar`Cookbookbat`,
-      },
+        avoid: sugarItemsAboutToBreak(),
+      }),
       choices: { 1387: 3 },
       limit: { tries: 1 },
     },
