+++ src/tasks/weapondamage.ts
@@ -24,6 +24,7 @@ import {
   SongBoom,
 } from "libram";
 import Macro from "../combat";
+import { sugarItemsAboutToBreak } from "../engine/outfit";
 import { Quest } from "../engine/task";
 import { advCost, CommunityServiceTests, logTestSetup, tryAcquiringEffect } from "../lib";
 
@@ -78,10 +79,11 @@ export const WeaponDamageQuest: Quest = {
           .trySkill($skill`Use the Force`)
           .abort()
       ),
-      outfit: {
+      outfit: () => ({
         weapon: $item`Fourth of May Cosplay Saber`,
         familiar: $familiar`Cookbookbat`,
-      },
+        avoid: sugarItemsAboutToBreak(),
+      }),
       choices: { 1387: 3 },
       limit: { tries: 1 },
     },
